Created own geometry widget to keep code simple and clean
Currently its proof of concept, wich may work not right.
The goal is, that mainwindow only have to call set(geometry) and the geometry widget detects and select the correct dropdown entry and fills in the data into the spin boxes.
Also  mainwindow only needs to call getGeometry() to get the geometryString for generating the map and save in profile settings