Try and make sscanf / snprintf work on both windows and Linux
sscanf_s exists in C11, but I didn't find it in CXX11
_snprintf_s does not exist in C11 or CXX11. C11 does have snprintf_s

In combination with the fact that I think sscanf and snprintf are not used
insecurely here (*), I used sscanf and snprintf, and added macros to select
the microsoft/C11 alternatives on Windows.

(*) Notes:
Any insecurity of sscanf does not apply, as it is not reading strings.

snprintf already does bounds-checking on the destination buffer, which is
is much to large anyway. And as compilers (VC++ too, I assume...) usually
check whether the format specifiers match the parameter types, there
should be no need for an extra-secure version here either.